DevTools: Provide 'WindowExceptionHandler' logging to the orchestration instead of crashing devtools
